Fix Jenkinsfile

+++ Jenkinsfile
@@ -19,30 +19,38 @@ pipeline {
                     sso_commands_upsert_img = sso_commands_upsert_registry + ":${env.BUILD_ID}"
                     println("${sso_commands_delete_img}")
                     println("${sso_commands_upsert_img}")
-                    ssoCommandsUpsertImage = docker.build("${sso_commands_upsert_img}", "-f ${env.WORKSPACE}/sso_commands/populate_index/Dockerfile .")
-                    ssoCommandsDeleteImage = docker.build("${sso_commands_delete_img}", "-f ${env.WORKSPACE}/sso_commands/delete_index/Dockerfile .")
+                    if (params.sso_commands == "Reset") {
+                        ssoCommandsUpsertImage = docker.build("${sso_commands_upsert_img}", "-f ${env.WORKSPACE}/sso_commands/populate_index/Dockerfile .")
+                        ssoCommandsDeleteImage = docker.build("${sso_commands_delete_img}", "-f ${env.WORKSPACE}/sso_commands/delete_index/Dockerfile .")
+                    }
                 }
             }
         }
 
         stage("Testing - running in Jenkins node") {
             steps {
-                powershell "docker run --name ${sso_commands_delete} ${sso_commands_delete_img}"
-                powershell "docker run --name ${sso_commands_upsert} ${sso_commands_upsert_img}"
+                if (params.sso_commands == "Reset") {
+                    powershell "docker run --name ${sso_commands_delete} ${sso_commands_delete_img}"
+                    powershell "docker run --name ${sso_commands_upsert} ${sso_commands_upsert_img}"
+                }
             }
         }
 
         stage("Stopping running container") {
             steps {
-                powershell "docker stop ${sso_commands_delete}"
-                powershell "docker stop ${sso_commands_upsert}"
+                if (params.sso_commands == "Reset") {
+                    powershell "docker stop ${sso_commands_delete}"
+                    powershell "docker stop ${sso_commands_upsert}"
+                }
             }
         }
 
         stage("Removing the container") {
             steps {
-                powershell "docker rm ${sso_commands_delete}"
-                powershell "docker rm ${sso_commands_upsert}"
+                if (params.sso_commands == "Reset") {
+                    powershell "docker rm ${sso_commands_delete}"
+                    powershell "docker rm ${sso_commands_upsert}"
+                }
             }
         }
     }
